requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-ratelimit==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
dj-database-url==2.1.0
sib-api-v3-sdk==7.6.0
fakeredis==2.31.0
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
types-redis==4.6.0.11
mypy==1.8.0
